Rename Double to Float

+++ stack-lang/sast/Constant.scala
@@ -3,5 +3,5 @@ package sast
 enum Constant:
   case Bool(value: scala.Boolean)
   case Int(value: scala.Int)
-  case Double(value: scala.Double)
+  case Float(value: scala.Double)
   case String(value: java.lang.String)
